+++ lib/hubbado/upsert_version.rb
@@ -1,4 +1,3 @@
-require "hubbado/upsert_version/version"
 require 'active_record'
 require 'configure'
 require 'mimic'