+++ ChangeLog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-# [Unrelased]
+# [3.0.0 - 2025-02-17]
 ## Changed
 - Upserted result is replaced with Updated and Inserted